refactor: update workflows notifier and file structure
